# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CMS Event Service",
    "description": "Receives CMS webhook events and serves the resulting copy of the CMS entities.\r\n\r\n**Authentication.** Every endpoint requires HTTP Basic credentials. Accounts and roles:\r\n\r\n| Role | Can do |\r\n|---|---|\r\n| `Cms` | `POST /cms/events` only. This is the CMS organization. |\r\n| `User` | Read published entities that were not disabled by an admin. |\r\n| `Admin` | Read everything, including unpublished and disabled entities, and disable or enable entities. |\r\n\r\n**Versions.** An entity only moves forward in version. `publish` makes it visible, `unPublish`\r\nkeeps its data but hides it, and `delete` removes it. An `unPublish` carries the payload, so a\r\nversion that was never published can still be stored.\r\n\r\n**Admin override.** Disabling an entity is a local overlay: it never changes CMS data, and later CMS\r\nevents do not undo it.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:7024",
      "description": "Local development"
    }
  ],
  "paths": {
    "/cms/events": {
      "post": {
        "tags": [
          "CmsEvents"
        ],
        "summary": "Ingest a batch of CMS events",
        "description": "Accepts up to 1000 events (publish, add, update, unPublish, delete). Each event is validated and applied on its own: the response lists the outcome of every event (Applied, Ignored as obsolete, Rejected as invalid, or Failed). Reserved for the CMS account.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CmsEvent"
                }
              },
              "example": [
                {
                  "type": "publish",
                  "id": "article-1",
                  "version": 3,
                  "payload": {
                    "title": "Travel guide"
                  },
                  "timestamp": "2024-01-02T10:00:00Z"
                },
                {
                  "type": "unPublish",
                  "id": "article-2",
                  "version": 4,
                  "payload": {
                    "title": "Draft"
                  },
                  "timestamp": "2024-01-02T11:00:00Z"
                },
                {
                  "type": "delete",
                  "id": "article-3",
                  "timestamp": "2024-01-02T12:00:00Z"
                }
              ]
            },
            "text/json": {
              "schema": {
                "type": "array"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/entities": {
      "get": {
        "tags": [
          "Entities"
        ],
        "summary": "List entities",
        "description": "Paged list ordered by id. Users see published, enabled entities; admins also see unpublished and disabled ones. page starts at 1; pageSize is 1-100.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfEntityResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfEntityResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfEntityResponse"
                }
              }
            }
          }
        }
      }
    },
    "/entities/{id}": {
      "get": {
        "tags": [
          "Entities"
        ],
        "summary": "Get one entity",
        "description": "Hidden entities are reported as 404 to regular users so their existence is not revealed.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EntityResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/entities/{id}/disable": {
      "post": {
        "tags": [
          "Entities"
        ],
        "summary": "Disable an entity (admin)",
        "description": "Hides the entity from regular users. A local override only: CMS data is not changed and later CMS events do not undo it.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/entities/{id}/enable": {
      "post": {
        "tags": [
          "Entities"
        ],
        "summary": "Enable an entity (admin)",
        "description": "Reverts a disable. Whether regular users see the entity still depends on whether the CMS has it published.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CmsEvent": {
        "required": [
          "type",
          "id",
          "timestamp"
        ],
        "type": "object",
        "properties": {
          "type": {
            "enum": [
              "publish",
              "add",
              "update",
              "unPublish",
              "delete"
            ],
            "type": "string",
            "description": "Case-insensitive. add and update are treated as publish."
          },
          "id": {
            "pattern": "^[A-Za-z0-9._:-]{1,100}$",
            "type": "string",
            "description": "Entity id in the CMS."
          },
          "version": {
            "minimum": 1,
            "type": "integer",
            "description": "Required for publish, add, update and unPublish; ignored for delete.",
            "format": "int32"
          },
          "payload": {
            "type": "object",
            "description": "Required for publish, add, update and unPublish; up to 262144 characters. An unPublish carries the payload of the version being unpublished."
          },
          "timestamp": {
            "type": "string",
            "description": "When the event happened in the CMS (ISO-8601).",
            "format": "date-time"
          }
        },
        "description": "One CMS event. Each event in a batch is validated on its own."
      },
      "BatchResult": {
        "required": [
          "results"
        ],
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventResult"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "applied": {
            "type": "integer",
            "format": "int32"
          },
          "ignored": {
            "type": "integer",
            "format": "int32"
          },
          "rejected": {
            "type": "integer",
            "format": "int32"
          },
          "failed": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "EntityResponse": {
        "required": [
          "id",
          "version",
          "payload",
          "isPublished",
          "isDisabledByAdmin",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "payload": {
            "$ref": "#/components/schemas/JsonElement"
          },
          "isPublished": {
            "type": "boolean"
          },
          "isDisabledByAdmin": {
            "type": "boolean"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "EventOutcome": {
        "type": "integer"
      },
      "EventResult": {
        "required": [
          "index",
          "id",
          "type",
          "outcome",
          "reason"
        ],
        "type": "object",
        "properties": {
          "index": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "outcome": {
            "$ref": "#/components/schemas/EventOutcome"
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "JsonElement": {},
      "PagedResultOfEntityResponse": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntityResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      }
    },
    "securitySchemes": {
      "Basic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
  "security": [
    {
      "Basic": []
    }
  ],
  "tags": [
    {
      "name": "CmsEvents"
    },
    {
      "name": "Entities"
    }
  ]
}